rm persistent_server.ts
