+++ src/www/app/app.ts
@@ -21,8 +21,7 @@ import {Server} from '../model/server';
 import {Clipboard} from './clipboard';
 import {EnvironmentVariables} from './environment';
 import {OutlineErrorReporter} from './error_reporter';
-import {OutlineServer} from './outline_server';
-import {OutlineServerRepository} from './persistent_server';
+import {OutlineServer, OutlineServerRepository} from './outline_server';
 import {Settings, SettingsKey} from './settings';
 import {Updater} from './updater';
 import {UrlInterceptor} from './url_interceptor';
